fix(core): Move generic types out from rocks, unexport types
